requirements.txt
PySide6>=6.6.0
requests>=2.31.0
python-dotenv>=1.0.1
pydantic>=2.7.0
